requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.10.0
google-adk
pandas
pydantic
pytest
python-dotenv
google-cloud-secret-manager
google-auth